+++ lib/index.dart
@@ -1,5 +1,9 @@
 // Export pages
-export '/login_signup/login_signup_widget.dart' show LoginSignupWidget;
+export '/signin_page/signin_page_widget.dart' show SigninPageWidget;
 export '/onboarding/onboarding_widget.dart' show OnboardingWidget;
-export '/profile_page/profile_page_widget.dart' show ProfilePageWidget;
-export '/base_page/base_page_widget.dart' show BasePageWidget;
+export '/profile_base_page/profile_base_page_widget.dart'
+    show ProfileBasePageWidget;
+export '/listing_base_page/listing_base_page_widget.dart'
+    show ListingBasePageWidget;
+export '/message_base_page/message_base_page_widget.dart'
+    show MessageBasePageWidget;
