+++ lib/listing_base_page/listing_base_page_model.dart
@@ -1,8 +1,8 @@
 import '/flutter_flow/flutter_flow_util.dart';
-import 'profile_page_widget.dart' show ProfilePageWidget;
+import 'listing_base_page_widget.dart' show ListingBasePageWidget;
 import 'package:flutter/material.dart';
 
-class ProfilePageModel extends FlutterFlowModel<ProfilePageWidget> {
+class ListingBasePageModel extends FlutterFlowModel<ListingBasePageWidget> {
   @override
   void initState(BuildContext context) {}
