+++ lib/profile_base_page/profile_base_page_model.dart
@@ -1,0 +1,11 @@
+import '/flutter_flow/flutter_flow_util.dart';
+import 'profile_base_page_widget.dart' show ProfileBasePageWidget;
+import 'package:flutter/material.dart';
+
+class ProfileBasePageModel extends FlutterFlowModel<ProfileBasePageWidget> {
+  @override
+  void initState(BuildContext context) {}
+
+  @override
+  void dispose() {}
+}
